Renamed and added `andTests.swift` to project
